introduce go client style config creating functions
